Reimplement _convert_message_part_to_str to avoid corruption
Previously we were using the email.Message.get_content() method.
This would cause the raw_data_manager to forcefully convert any
payload with a Content-Type of text/* to a string.

This behaviour is unwelcome with GP2GP Large Message spec
where an attachment can claim to be a "text/plain" for example, but
in reality contain the GZIP compressed base64 representation of
"text/plain".

There was previously code to workaround the raw_data_manager,
however the workaround was buggy in the case that the attachment was
encoded with base64 but not decompressable (for example because it's
a fragment).

This new implementation avoids the ContentManager by calling the
email.Message.get_payload() instead.

Also avoid guessing whether a payload was base64 or not by inspecting
the Content-Transfer-Encoding header directly. This should make the
behaviour of the inbound adaptor more predictable, and also matches
how the `email` package decides whether a message is base64 or not.

https://docs.python.org/3/library/email.contentmanager.html#email.contentmanager.raw_data_manager