Remove the Docker base images
There appears to be a 1-2-1 mapping between actual image, and base
image.

This layer of indirection doesn't appear to be adding any value to
the MHS Adaptor, and in fact makes it harder to release as there
are two separate images for each of the inbound,outbound,sds.

By combining the base image into the actual image this will
simplify maintenance for the developers.