Outbound: Specify the OpenSSL Root CA Bundle when making SDS requests (#191)
When testing the Adaptor in PTL, we noticed that the first Outbound request would return
a successful response, and then all subsequent requests returned a 500. We narrowed the
issue down to the tornado HTTP client reusing the last CA Bundle it was given if no CA Bundle
was explicitly specified. After the first request, this meant that SDS requests were trying to
validate the authenticity of the SDS API by using the NHS root certificates (because the
last request made was to Spine).

To workaround this Tornado issue, this we've changed our CommonHttps.make_request
method to explicitly specify the OpenSSL root CA bundle in the case that no bundle is
provided. This should only affect the calls to the FHIR SDS API, as that's the only HTTP
server we call that uses a standard certificate. Calls to Spine use the NHS certificates
passed in as environment variables.

---------

Co-authored-by: Adrian Clay <adrian@madetech.com>