Switch from using async_test to IsolatedAsyncioTestCase (#204)
IsolatedAsyncioTestCase was introduced within Python 3.8, which was
after the original creation of these tests.

By replacing the custom async_test implementation with
IsolatedAsyncioTestCase it will make our tests more idiomatic and
therefore more maintainable.

Additionally, the implementation of `async_test` uses asyncio.coroutine,
a method which was deprecated in Python 3.8 and will be removed in Python
3.11